Class tree refactoring (Entity -> Entity.Continuous) snapshot
