+++ AngouriMath/Core/Entity/Omni/Entity.Set.cs
@@ -21,6 +21,9 @@ namespace AngouriMath
 {
     partial record Entity
     {
+        public partial record Set
+        {
 
+        }
     }
 }
